Improving experiments and testing updated skaworkflows generation.

+++ chapter3/initial_results/run_comparisons.py
@@ -44,26 +44,30 @@ def run_shadow(tup, queue, test=False):
     if test:
         print(f"{tup}")
         return None
-    wf, env, f, graph, telescope, position, nodes, channels, lock = tup
-    if "no_data" in wf.name:
-        data = False
+    wf, env, f, graph, telescope, position, nodes, channels, lock, pipeline_sets = tup
+
+    if "standard" in wf.name:
+        data = 'no_data'
+    elif "no_data" in  wf.name and "edges" in wf.name:
+        data = 'edges'
     else:
-        data = True
+        data = 'task_edges'
+
     workflow = Workflow(wf)
     hpso = f.split("_")[0]   
 
     workflow.add_environment(env)
     print(f"Running FCFS {position} {wf.name}")
-    # fcfs_res = fcfs(workflow, position).makespan
-    # # heft_res = None
-    # res_str_fcfs = (f"{hpso},workflow,{fcfs_res},"
-    #                 f"{graph},{telescope},{nodes},{channels},{data}")
-    # print(res_str_fcfs)
-    # lock.acquire()
-    # with output.open('a') as f:
-    #     f.write(f"{str(res_str_fcfs)}\n")
-    #     f.flush()
-    # lock.release()
+    fcfs_res = fcfs(workflow, position).makespan
+    # heft_res = None
+    res_str_fcfs = (f"{hpso},workflow,{fcfs_res},"
+                    f"{graph},{telescope},{nodes},{channels},{data},{pipeline_sets}")
+    print(res_str_fcfs)
+    lock.acquire()
+    with output.open('a') as f:
+        f.write(f"{str(res_str_fcfs)}\n")
+        f.flush()
+    lock.release()
 
 
 def run_parametric(tup, queue, test=False):
@@ -84,8 +88,8 @@ def run_parametric(tup, queue, test=False):
     """
     if test:
         return None
-
-    wf, env, f, graph, telescope, position, nodes, channels, lock = tup
+    # print(tup)
+    wf, env, f, graph, telescope, position, nodes, channels, lock, pipeline_sets = tup
     if graph == 'scatter' or channels == 512 or "no_data" not in wf.name:
         return None
     if "no_data" in wf.name:
@@ -93,8 +97,11 @@ def run_parametric(tup, queue, test=False):
     else:
         data = True
     hpso = f.split("_")[0]
+    pipeline_list = None
+    if pipeline_sets:
+        pipeline_list = pipeline_sets.split('_')
     par_res = calculate_parametric_runtime_estimates(
-        PAR_MODEL_SIZING, telescope, [hpso]
+        PAR_MODEL_SIZING, telescope, [hpso], pipeline_list
     )
     res_str = (
         f"{hpso},parametric,{par_res[hpso]['time']},"
@@ -116,87 +123,90 @@ def run_test(tup, queue, test=True):
     lock.release()
 
 
-def listener(queue, output: Path):
-    """Listen for messages on the queue and write updates to the XML log."""
-
-    with output.open('a') as f:
-        while True:
-            msg = queue.get()
-            if str(msg) == 'Finish':
-                # print(msg)
-                break
-            f.write(str(msg))
-            f.flush()
-
-
-def low_setup(config_iterations, channel_iterations, wfdict, lock):
+def low_setup(config_iterations, channel_iterations, wfdict, lock, pipeline_sets):
     print('Low Setup')
     data_iterations = [False] # , True]
-    graph_iterations = ['prototype', 'scatter']
+    graph_iterations = ["prototype"] #, 'scatter']
     count = 0
     config = 896
-    hpsos=['hpso01','hpso02a', 'hpso2b']
+    hpsos=['hpso01','hpso02a', 'hpso02b']
     for hpso in hpsos:
         for data in data_iterations:
             for channel in channel_iterations:
                # Set up paths to the respective directory
                 for graph in graph_iterations:
                     low_path = Path(
-                        f"{BASE_DIR}/low/{graph}/"
+                        f"{BASE_DIR}/low_maximal/{graph}/"
                     )
                     low_config_shadow = config_to_shadow(low_path / f"no_data_low_sdp_config_{graph}_n{config}_{channel}channels.json", 'shadow_')
                     low_env = Environment(low_config_shadow)
-                    f = low_path / 'workflows' / f"{hpso}_time-18000_channels-{channel}_tel-512_no_data.json"
-                    fdata = low_path / 'workflows' / f"{hpso}_time-18000_channels_{channel}_tel-512.json"
-                    nenv = deepcopy(low_env)
-                    for x in range(10):
+                    f = low_path / 'workflows' / f"{hpso}_time-18000_channels-{channel}_tel-512_no_data-standard.json"
+                    fedges = low_path / 'workflows' / f"{hpso}_time-18000_channels-{channel}_tel-512_no_data-edges.json"
+                    fdata = low_path / 'workflows' / f"{hpso}_time-18000_channels-{channel}_tel-512-edges.json"
+                    for x in range(1):
+                        nenv = deepcopy(low_env)
                         wfstr = f"{f.name}_{graph}_{x}"
+                        wfstr_edges = f"{fedges.name}_{graph}_{x}"
                         wfstr_data = f"{fdata.name}_{graph}_{x}"
                         wfdict[wfstr] = (
                             f, nenv, f.name, graph, 'low-adjusted', count,
-                        config, channel,lock
+                        config, channel, lock, pipeline_sets
                         )
                         wfdict[wfstr_data] = (
                             fdata, nenv, fdata.name, graph, 'low-adjusted', count,
-                        config, channel,lock
+                        config, channel, lock, pipeline_sets
+                        )
+                        wfdict[wfstr_edges] = (
+                            fedges, nenv, fedges.name, graph, 'low-adjusted', count,
+                        config, channel, lock, pipeline_sets
                         )
                         if graph == 'scatter':
                             break
                         count += 1
     return wfdict
 
-def mid_setup(config_iterations, channel_iterations,wfdict,lock):
+def mid_setup(config_iterations, channel_iterations,wfdict,lock, pipeline_sets):
     data_iterations = [False, True]
-    graph_iterations = ['prototype', 'scatter']
+    graph_iterations = ['prototype'] #, 'cont_img_mvp_graph']
     count = 0
-    config = 896
-    for data in data_iterations:
-        for channel in channel_iterations:
-            # Set up paths to the respective directory
-            if config == 512 and channel == 896:
-                # We are not interested in this experiment
-                continue
-            for graph in graph_iterations:
-                mid_path = Path(
-                    f"{BASE_DIR}/mid/{graph}/"
-                )
-                mid_config_shadow = config_to_shadow(mid_path / f"no_data_mid_sdp_config_{graph}_n{config}_{channel}channels.json", 'shadow_')
-                mid_env = Environment(mid_config_shadow)
-                for f in (Path(mid_path) / 'workflows').iterdir():
+    config = 786
+    hpsos=['hpso13','hpso15', 'hpso22', 'hpso32']
+    hpso_times = {'hpso13': 28800 ,'hpso15':15840 , 'hpso22':28800, 'hpso32':7920}
+    for hpso in hpsos:
+        for data in data_iterations:
+            for channel in channel_iterations:
+               # Set up paths to the respective directory
+                for graph in graph_iterations:
+                    mid_path = Path(
+                        f"{BASE_DIR}/mid_maximal/{graph}/"
+                    )
+                    mid_config_shadow = config_to_shadow(mid_path / f"no_data_mid_sdp_config_{graph}_n{config}_{channel}channels.json", 'shadow_')
+                    mid_env = Environment(mid_config_shadow)
+                    f = mid_path / 'workflows' / f"{hpso}_time-{hpso_times[hpso]}_channels-{channel}_tel-197_no_data-standard.json"
+                    fedges = mid_path / 'workflows' / f"{hpso}_time-{hpso_times[hpso]}_channels-{channel}_tel-197_no_data-edges.json"
+                    fdata = mid_path / 'workflows' / f"{hpso}_time-{hpso_times[hpso]}_channels-{channel}_tel-197-edges.json"
                     nenv = deepcopy(mid_env)
-                    wfstr = (
-                        f"{f.name}{graph}mid-adjusted{config}"
-                        f"{channel}")
-                    if wfstr in wfdict:
-                        continue
-                    else:
+                    for x in range(1):
+                        wfstr = f"{f.name}_{graph}_{x}"
+                        wfstr_edges = f"{fedges.name}_{graph}_{x}"
+                        wfstr_data = f"{fdata.name}_{graph}_{x}"
                         wfdict[wfstr] = (
-                        f, nenv, f.name, graph, 'mid-adjusted', count,
-                        config, channel,lock)
-                    count += 1
+                            f, nenv, f.name, graph, 'mid-adjusted', count,
+                        config, channel, lock, pipeline_sets
+                        )
+                        wfdict[wfstr_data] = (
+                            fdata, nenv, fdata.name, graph, 'mid-adjusted', count,
+                        config, channel,lock, pipeline_sets
+                        )
+                        wfdict[wfstr_edges] = (
+                            fedges, nenv, fedges.name, graph, 'mid-adjusted', count,
+                        config, channel,lock, pipeline_sets
+                        )
 
+                        if graph == 'scatter':
+                            break
+                        count += 1
     return wfdict
-
 
 if __name__ == '__main__':
     BASE_DIR = Path(f"/home/rwb/Dropbox/University/PhD/experiment_data/chapter3/initial_results/")
@@ -206,20 +216,23 @@ if __name__ == '__main__':
     manager = Manager()
     queue = manager.Queue()
     lock = manager.Lock()
+    pipeline_sets = None # None = Maximal use case with all workflows
 
-    wfs_dict = low_setup(low_config_iterations, low_channel_iterations, wfs_dict, lock)
-    # mid_config_iterations = [786] #, 512]
-    # mid_channel_iterations = [896] # , 512]
-    # wfs_dict = mid_setup(mid_config_iterations, mid_channel_iterations, wfs_dict, lock)
+    wfs_dict = low_setup(low_config_iterations, low_channel_iterations, wfs_dict, lock, pipeline_sets)
+    mid_config_iterations = [786] #, 512]
+    mid_channel_iterations = [786] # , 512]
+    wfs_dict = mid_setup(mid_config_iterations, mid_channel_iterations, wfs_dict, lock, pipeline_sets)
     # for e in sorted(wfs_dict.keys()):
     #     print(e)
     output = Path(
         f"chapter3/initial_results/results_{date.today().isoformat()}.csv")
-    with output.open('w+') as fo:
-        fo.write('hpso,workflow,time,graph,telescope,nodes,channels,data')
+    if not output.exists:
+        with output.open('w+') as fo:
+            fo.write('hpso,workflow,time,graph,telescope,nodes,channels,data,pipeline_sets')
 
     params = set(zip(wfs_dict.values(), [queue for x in range(len(wfs_dict))]))
-    # with Pool(processes=1) as pool:
-    #     result = pool.starmap(run_parametric, params)
+
+    with Pool(processes=1) as pool:
+        result = pool.starmap(run_parametric, params)
     with Pool(processes=4) as pool:
         result = pool.starmap(run_shadow, params)
